frontend/requirements.txt
streamlit
httpx
plotly
python-dotenv
tenacity
streamlit-autorefresh
